pyproject.toml
[project]
name = "pufferpantry"
version = "0.1.0"
description = "A meal planning app for the family — powered by pufferfish energy"
requires-python = ">=3.10"

dependencies = [
    # Web framework
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "python-multipart>=0.0.9",

    # Database
    "sqlalchemy>=2.0",
    "alembic>=1.13",

    # Configuration
    "pydantic-settings>=2.1",

    # Claude API
    "anthropic>=0.40",

    # Google APIs
    "google-auth>=2.27",
    "google-auth-oauthlib>=1.2",
    "google-api-python-client>=2.114",

    # Image handling
    "pillow>=10.2",

    # HTTP client (for USDA nutrition API)
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
    "ruff>=0.2",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
